+++ 2024-02-13_Handin
@@ -1,0 +1,5 @@
+Progress so far:
+Gavin and Elaine finished SQL for the project.
+Elaine has started on the frontend and backend of Part 3. The draft was roughly finished except several unsure uses.
+
+Future planning:
