+++ Group.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Find a Group</title>
+</head>
+<body>
+<div>Search for group member(s)</div>
+<form method="post">
+    <label for="fName">First Name:</label><br>
+    <input type="fName" id="fName" name="fName"><br>
+    <label for="lName">Last Name:</label><br>
+    <input type="lName" id="lName" name="lName"><br>
+    <label for="phone">Phone Number:</label><br>
+    <input type="phone" id="phone" name="phone"><br>
+    <input type="submit" value="Search">
+</form>
+</body>
+</html>
