+++ Login-Student.html
@@ -5,13 +5,13 @@
     <title>Log In</title>
 </head>
 <body>
-    <a href="/Signup.html" class="btn btn-primary" type="button">Sign Up</a>
+    <a href="/Signup-Student.html">Sign Up</a>
     <form method="post">
         <label for="phone">Phone Number: </label><br>
         <input type="text" id="phone" name="phone"><br>
         <label for="password">Password: </label><br>
         <input type="password" id="password" name="password"><br><br>
-        <button type="button" onclick="login()">Log in</button>
+        <input type="submit" name="Log in">
     </form>
     <script src="Login-Student.js"></script>
 </body>
