+++ Signup-Student.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Sign Up</title>
-    <a href="/Login-Student.html" class="btn btn-primary" type="button">Log in</a>
+    <a href="/Login-Student.html">Log in</a>
     <form method="post">
         <label for="phone">Phone Number: </label><br>
         <input type="text" id="phone" name="phone"><br>
@@ -13,7 +13,7 @@
         <input type="fName" id="fName" name="fName"><br>
         <label for="lName">Last Name:</label><br>
         <input type="lName" id="lName" name="lName"><br><br>
-        <button type="button" onclick="signup()">Sign up</button>
+        <input type="submit" value="Sign up">
     </form>
 </head>
 <body>
