+++ Student-Elaine.html
@@ -8,7 +8,8 @@
     <h2>Student Information</h2>
     <div><strong>Profile: </strong></div>
     <script>
-        
+
     </script>
+    <a href="/Group.html">Find a Group</a>
 </body>
 </html>
