+++ index.html
@@ -5,7 +5,7 @@
     <title>Housing Project</title>
 </head>
 <body>
-    <a href="/Login-Student.html" class="btn btn-primary" type="button">Students Log In</a><br>
-    <a href="/Signup.html" class="btn btn-primary" type="button">Students Sign Up</a><br>
+    <a href="/Login-Student.html">Students Log In</a><br>
+    <a href="/Signup-Student.html">Students Sign Up</a><br>
 </body>
 </html>
